+++ sklearn/tree/_export.py
@@ -832,7 +832,7 @@ def export_text(decision_tree, feature_names=None, max_depth=10,
 
     >>> from sklearn.datasets import load_iris
     >>> from sklearn.tree import DecisionTreeClassifier
-    >>> from sklearn.tree.export import export_text
+    >>> from sklearn.tree import export_text
     >>> iris = load_iris()
     >>> X = iris['data']
     >>> y = iris['target']
