+++ sklearn/tree/tests/test_tree.py
@@ -38,8 +38,8 @@ from sklearn.tree import ExtraTreeRegressor
 
 from sklearn import tree
 from sklearn.tree._tree import TREE_LEAF, TREE_UNDEFINED
-from sklearn.tree.tree import CRITERIA_CLF
-from sklearn.tree.tree import CRITERIA_REG
+from sklearn.tree._tree_base import CRITERIA_CLF
+from sklearn.tree._tree_base import CRITERIA_REG
 from sklearn import datasets
 
 from sklearn.utils import compute_sample_weight
